Move structopt structs to src/cli/syntax
